refactor(operations): extract static methods from Deployment
