+++ operations/src/deployment/deployment.js
@@ -4,6 +4,10 @@ const { join } = require('node:path')
 const { writeFile: write } = require('node:fs/promises')
 const { yaml, directory: { copy } } = require('@toa.io/gears')
 
+const { merge } = require('./deployment/merge')
+const { declare } = require('./deployment/declare')
+const { describe } = require('./deployment/describe')
+
 /**
  * @implements {toa.operations.deployment.Deployment}
  */
@@ -24,10 +28,10 @@ class Deployment {
    * @param process {toa.operations.Process}
    */
   constructor (context, compositions, dependencies, process) {
-    const dependency = Deployment.#merge(dependencies)
+    const dependency = merge(dependencies)
 
-    this.#declaration = Deployment.#declare(context, dependency)
-    this.#contents = Deployment.#describe(compositions, dependency)
+    this.#declaration = declare(context, dependency)
+    this.#contents = describe(compositions, dependency)
     this.#process = process
   }
 
@@ -70,80 +74,6 @@ class Deployment {
       ['template', this.#declaration.name, ...args, this.#target],
       { silently: true })
   }
-
-  /**
-   * @param dependencies {Array<toa.operations.deployment.Dependency>}
-   * @returns {toa.operations.deployment.Dependency}
-   */
-  static #merge (dependencies) {
-    /** @type {Array<toa.operations.deployment.Reference>} */
-    const references = []
-    /** @type {Array<toa.operations.deployment.Service>} */
-    const services = []
-
-    for (const dependency of dependencies) {
-      if (dependency.references !== undefined) references.push(...dependency.references)
-      if (dependency.services !== undefined) services.push(...dependency.services)
-    }
-
-    return { references, services }
-  }
-
-  /**
-   * @param context {toa.formation.Context}
-   * @param dependency {toa.operations.deployment.Dependency}
-   * @returns {toa.operations.deployment.Declaration}
-   */
-  static #declare (context, { references }) {
-    const { name, description, version } = context
-
-    const dependencies = references.map(({ values, ...rest }) => rest)
-
-    return {
-      ...DECLARATION,
-      name,
-      description,
-      version,
-      appVersion: version,
-      dependencies
-    }
-  }
-
-  /**
-   * @param compositions {Array<toa.operations.deployment.Composition>}
-   * @param dependency {toa.operations.deployment.Dependency}
-   * @returns {toa.operations.deployment.Contents}
-   */
-  static #describe (compositions, { references, services }) {
-    /** @type {Set<string>} */
-    const components = new Set()
-
-    for (const composition of compositions) {
-      for (const component of composition.components) {
-        components.add(component)
-      }
-    }
-
-    const dependencies = references?.reduce((map, reference) => {
-      const { name, alias, values } = reference
-
-      map[alias || name] = values
-
-      return map
-    }, {})
-
-    return {
-      compositions,
-      components: Array.from(components),
-      services,
-      ...dependencies
-    }
-  }
-}
-
-const DECLARATION = {
-  apiVersion: 'v2',
-  type: 'application'
 }
 
 const TEMPLATES = join(__dirname, 'chart/templates')
